Fix ::backdrop user-agent rule for popovers
With the new css-position spec and overlay CSS property, it is possible
for ::backdrop and its corresponding element with the popover attribute
to be in the top layer while it is transitioning from showing to hidden.
During this time, :popover-open does not apply but the element and
::backdrop are still in the top layer, but we still want our
popover-specific user-agent style rules for ::backdrop to apply. This PR
accomplishes this by moving the ::backdrop style rule from a stylesheet
to a more specific definition.

This was implemented in chromium by replacing the
:popover-open::backdrop selector with an internal pseudo-selector.