Ignore DOM state when hiding popovers if needed
Fixes https://github.com/whatwg/html/issues/9161
Fixes https://github.com/whatwg/html/issues/9367
Makes obsolete https://github.com/whatwg/dom/pull/1185

This PR prevents the hide popover algorithm from returning early when
the popover attribute is removed or when the element with the popover
attribute is removed from the document.

The fireEvents parameter is used as an indicator that either the element
is being removed or that the attribute is being removed, and when it is
false, the calls to check popover validity are replaced with a check to
simply see if the popover is already hidden.

This patch also makes removal of the popover attribute stop firing
events in order to signal to the hide popover algorithm that checks for
the popover attribute should be ignored.