Add customElements.getName(constructor)
In CustomElements v0 you could use the `.name` field to get the defined
tag name, but CustomElements v1 does not offer such a field. `localName`
can be used within an instance but this prohibits use in, for example,
static methods.

This was discussed in the WCCG F2F, tracked in
https://github.com/WICG/webcomponents/issues/566. The conclusion of the
F2F was to write up a spec change and WPT tests.