Editorial: fix popover attribute index entries
Fix the backlinks to the attribute definitions, which has the effect of causing the attribute definitions to show up in the input and button element definitions. (The newlines here are, unfortunately, important.)

Also make popovertarget use the same "ID*" value space as other ID-accepting elements, as noted in https://github.com/whatwg/html/pull/9144#discussion_r1244594592.