Allow img element to be a child of picture element
Previously picture element specified that it contains source elements
and one img element, but only source element specified that it can
appear as a child of picture element. Now img element also explicitly
allows this.

This is especially useful for automated tooling which might scrape
through the document and expect a parent-child relationship to be
mutual.