Editorial: clarify adoption agency step 13.3
After implementing the adoption agency algorithm as described in
the specification, I was very confused why some respective tests
of html5lib-tests were failing. As it turns out I misinterpreted:

> If node is formatting element, then break.

as "if node is a formatting element" (i.e. any element in the
formatting category). I missed that "formatting element" referred
to the "formatting element" variable. Indeed, the source file does
declare it with the <var> tag, however in the rendered spec this
just becomes italic text, which evidently is easy to miss.

This commit therefore clarifies the wording to:

> If node is the formatting element, then break.