Editorial: use camelCase for variables in adoption agency algorithm
The motivation for this change is that previously:

> If node is formatting element, then break.

could easily be misunderstood as "if node is a formatting element"
(i.e. any element in the formatting category), instead of referring
to the "formatting element" variable. While the source file does
declare the variables with the <var> tag, in the rendered spec
this just becomes italic text, which is easy to miss.

This commit therefore changes all the variables in the adoption agency
algorithm to use camelCase, to prevent such possible confusion.

Trivia: camelCase is apparently the more modern style in WHATWG specs.
Before this change the spec contained ~1,500 var elements containing
a space and ~5,000 var elements using camelCase, as reported by:

    [...document.querySelectorAll('var')].reduce(
      ([spaceCase, camelCase], v) => [
        spaceCase + v.textContent.trim().includes(' '),
        camelCase + (v.textContent != v.textContent.toLowerCase()),
      ],
      [0, 0]
    );