Add commandfor & command attributes to HTMLButtonElement
This adds the commandfor & command attributes and a "command" event using the CommandEvent interface.

Button activation checks if the button has a "commandfor" target and if so performs invoker command
behaviour depending on command and the target element.

This introduces 1 new IDLs:

- CommandEvent: A new event that has `action` and `source`.

And extends the HTMLButtonElement IDL to introduce `command` and
`commandfor`

Things NOT covered in this commit that are included in the proposal (https://github.com/whatwg/html/issues/9625):

- Default per element behaviours for anything beyond popover & dialog,
  this will be dealt with in subsequent individual commits.

- `interestaction` and `interesttarget`.