Make computation of directionality account for shadow trees
This specifies (with some additional detail) the proposal in https://github.com/whatwg/html/issues/3699#issuecomment-951423468. This changes three things:

* the inheritance of directionality,
* the inheritance of language, and
* the computation of dir=auto to account for shadow trees.

This builds on the work in #9452 and #9554 to refactor this section, and builds on work by Brian Kardell, Eric Meyer, and others in #7424 and in #9166 and work by fantasai, rniwa, smaug, MyIdShin, Brian Kardell, and others in #3699.

Fixes #3699.