Fix error (resulting from confusing text) in auto directionality traversal.
This fixes an unintended behavior in the auto directionality traversal
that I wrote in 11dc4c7557f5a4c879c1ce65b9cb59b91310bbf5 (PR #9796,
fixing #3699) by failing to edit existing text enough.

This section specifies an algorithm that implementors are expected to
turn into a traversal of a subtree that skips parts of the subtree.
However, by describing the skipping test as examining ancestors rather
than ancestors-or-self, it introduces a subtle difference between the
algorithm that an implementor is likely to write and what the spec
describes.  This change changes the test to ancestors-or-self.

My intent when writing this was that a <slot> element in a shadow tree
with a dir attribute would be skipped by the traversal, but what I
actually wrote was the opposite, specifying that it would provide the
result of its dir attribute as the first-strong directionality.  (That
was my intent because it was slightly easier to implement, I needed to
pick one way or the other, and I didn't see any other reasons to pick
one or the other.)