Fix dir=auto handling of descendant <slot> elements.
In 11dc4c7557f5a4c879c1ce65b9cb59b91310bbf5 (PR #9796, fixing #3699) I
was trying to specify the proposal in
https://github.com/whatwg/html/issues/3699#issuecomment-951423468 .  I
found a case where I had intended to match that proposal but made an
error: the handling of <slot> descendants was intended to look at their
resolved directionality, not their auto directionality.  (And this is a
simpler test, since the concept always exists, so there's no need to
test it for null.)  This corrects that error.