Make language of a node consider lang attribute on SVG elements.
The rules for finding the language of a node were made somewhat more
formal in 11dc4c7557f5a4c879c1ce65b9cb59b91310bbf5 (PR #9796,
fixing #3699).  Both before and after that change they considered the
lang attribute in the XML namespace, and the lang attribute in no
namespace on HTML elements.  However, the SVG spec at
https://www.w3.org/TR/2018/CR-SVG2-20181004/struct.html#LangSpaceAttrs
has very similar wording to the HTML spec prior to that edit, and also
defined a lang attribute on SVG elements.

This updates the language of a node algorithm in HTML to consider SVG
elements as well, since implementing it as written (without this change)
breaks the lang attribute in SVG.