Defer persisted state restore when view transition is started
When starting a view transition right before state restoration,
e.g. `addEventListener("popstate", () => startViewTransition(() => {...}))`,
according to the current spec the scroll position would be restored
before the `startViewTransition` callback is called, which means that
it would be captured as part of the old state, even though it belongs
to the new state.

This change makes it so that we defer persisted state restoration
until after the old state is captured.

Together with https://github.com/w3c/csswg-drafts/pull/9676