Fix case where the directionality of a <bdi> could be null.
In #9796 I refactored the term "auto directionality" to move the
fallback to the parent directionality out of the definition of "auto
directionality", since I needed the null result in one case to continue
a tree traversal rather than returning.

However, I missed fixing up one of the references to the term (should I
blame the line break?), leaving a case where the spec could define the
directionality of a <bdi> element as null.

This fixes that case, and converts the <ol> to a <dl class="switch">.