Hide all popovers when beforetoggle shows a popover
From @mfreed7's notes in http://crrev.com/c/5141430

It was previously possible to hit some underspecified behavior with
something like this:

```html
<div popover id=p1>Popover 1
  <div popover id=p2>Popover 2</div>
</div>
<script>
  p1.showPopover();
  p1.addEventListener('beforetoggle',() => p2.showPopover());
  p1.hidePopover();
</script>
```

The problem is that "hide all popovers until" doesn't end up with
the desired "until" popover on the top of the stack in this case.
There is already a similar situation within the "hide all..."
algorithm itself, but that only handles the case where a popover
being hidden by "hide all..." has the beforetoggle listener.
This is the same problem, but for the case that the "until" popover
has that listener.