Add a close event to MessagePort
As discussed in #1766, there is currently no timely and reliable way to detect when a MessagePort becomes disentangled. This makes it difficult to release resources associated with ports.

This adds a close event to MessagePort objects. For example, if there are two entangled ports, portA and portB, and portA is closed, the close event is fired on portB. See https://github.com/fergald/explainer-messageport-close for an explainer.

Closes #1766.