Refactor event loop processing model to perform rendering as a task
Instead of "updating the rendering" after each task, we wait for rendering opportunities in parallel, and post a special task to update the rendering. This is more in line with existing implementations, and fixes weird cases where e.g. a microtask queue can run without a current task.

Helps with #9824.