Drop duplicate dfns of width and height IDL attr for HTMLImageElement
The `width` IDL attribute for `HTMLImageElement` is correctly defined next to the definition of the `img` element:
https://html.spec.whatwg.org/multipage/embedded-content.html#dom-img-width

Same thing for `height`.

Section 4.8.17 Dimension attributes also talks about `width` and `height` IDL attributes for *other* elements with dimensions:
https://html.spec.whatwg.org/multipage/embedded-content-other.html#dom-dim-width

The prose itself is correct, but the `for` attribute of the `dfn` elements also listed `HTMLImageElement`, thus creating a second definition of these IDL attributes for `HTMLImageElement`. This update fixes the `dfn` elements.