Add dialog associated invocation action algorithm
This specifies what buttons with an invoketarget pointing to a dialog element
should do, based on the Invokers proposal (https://github.com/whatwg/html/issues/9625).

This introduces new specifications within the HTMLDialogElement, giving it an "associated invocation action algorithm" such that:

 - If an `invoketarget` points to an HTMLDialogElement element
    - If the `invokeaction` is `auto` or `close`, and the Dialog is
      open, then try to close the dialog.
    - If the `invokeaction` is `auto` or `showmodal` and the Dialog is
      not open, then try to open the dialog.

If the `invokeaction` is none of the above, then it will do nothing.

Things NOT covered in this commit that are included in the proposal (https://github.com/whatwg/html/issues/9625):

- Other per element behaviours (invocation action algorithms) beyond
  HTMLDialogElement. These will be dealt with in subsequent commits
  adding each elements "invocation action algorithm".

- `interestaction` and `interesttarget`.