Improve cross-document deactivation algorithm
This refactors the cross-document deactivation part of the navigation algorithm to use fewer closures and instead a couple of top-level algorithms.

It also fixes an issue where the incorrect document was being checked for the view transitions opt-in. This integrates with the new View Transitions hook being introduced in https://github.com/w3c/csswg-drafts/pull/10078.