Allow top layer elements to be nested within popovers
This allows top layer elements, including the dialog element, to be nested inside of an open popover, by not closing the popover when the new top layer element is opened. Without this patch, opening a modal dialog inside of a popover will make the page inert and make the dialog invisible.

Fixes #9998. See also https://github.com/whatwg/fullscreen/pull/237.