Further fixes to keydown vs. keyup for close requests
The original text was written based on the misunderstanding that keyup was the correct event. A small patch was made in d785cbe995efd86d885c36eb270a10dc9400fa71, but it was not complete; other parts of the text were still wrong, or overly-complex. This rewrites the relevant requirements and examples to be clearer, now that we know keydown is correct.

Closes #10523.