Overhaul exception reporting algorithms
This carries out the first portion of the plan in https://github.com/whatwg/html/issues/958#issuecomment-226309871, by creating a new "report an exception" algorithm that replaces the previous "report an error" and "report the exception".

The new algorithm directly includes the error propagation and fallback behavior, and requires callers to supply the global object to be used, rather than magically inferring it. It no longer takes a script, since in most cases the script was not rigorously determined. Follow-up work on determining the correct muting behavior (which the script argument was used for) is tracked in #10514.

All call sites within HTML are updated. #10516 tracks updating call sites in other specifications. In most cases the global used for HTML is now specified rigorously and matching implementations. #10526 and #10527 track the remaining cases with interop issues.

Closes #958, with the rest of the plan there tracked via the issues mentioned above, https://github.com/whatwg/webidl/issues/1423, and the https://github.com/whatwg/html/labels/topic%3A%20error%20reporting label.