Reconstruct active formatting elements for PLAINTEXT element.
Resolves #8009

All major HTML parsers reconstruct active formatting elements when
inserting a new PLAINTEXT element, leaving formatting elements as
children of the PLAINTEXT element.

However, the spec implies that this should not happen, because it
doesn't instruct reconstruction. The implication in the spec is that
a PLAINTEXT element may contain no children other than the plaintext
content of the remainder of the HTML document.

> Once a start tag with the tag name "plaintext" has been seen, that
> will be the last token ever seen other than character tokens
> (and the end-of-file token), because there is no way to switch out
> of the PLAINTEXT state.

This patch updates the spec to conform to the existing implementations
by adding the mention to trigger reconstruction.