Fix incorrect recursion for dir=auto special cases.
dir=auto has special-case behavior for both (1) auto-directionality
form-associated elements and (2) slots with assigned nodes.  In these
special cases dir=auto is intended to have special behavior when used on
the element, but this special behavior was not intended to apply when
dir=auto was used on an ancestor.

The wording in the HTML spec for this behavior got this incorrect for at
least one case, where the handling of the slot case needs recursion over
the assigned nodes as though they were descendants.  This was uncovered
by the Gecko implementor adding a test for this case as described in
https://bugzilla.mozilla.org/show_bug.cgi?id=1876163#c13 .  I believe
this was a mistake in the spec and should be fixed.