WebAssembly JS module integration including source phase imports
Integrate the https://github.com/WebAssembly/esm-integration specification with HTML, including support for source phase imports (https://github.com/tc39/proposal-source-phase-imports) and synchronous instantiation.

There is an early validation error algorithm that is applied when modules are created, which needed to be updated to support the source phase imports proposal. Specifically, constructed modules may not have their dependencies loaded when in the source phase. This validation logic is now located in HostLoadImportedModule, and performed against all referrer dependencies on the first call to HostLoadImportedModule for a given referrer, detected by the first argument matching the first module of the Cyclic Module Record (indicating the module is loading its dependencies).