Apply globalAlpha on the canvas filter source
The current specification for canvas filters applies the globalAlpha
on the filter result. The drawing model essentially says:
 - Render the shape, creating image A.
 - When the current filter is set, use it on image A to create image B.
 - Multiply the alpha component of every pixel in B by global alpha.

This leads to a unexpected behavior. When drawing a transparent shape
with a drop-shadow filter, the shadow should be visible through the
transparent foreground. This should be true whether the drawing is made
transparent via fillStyle or via globalAlpha. This is how Chrome and
Firefox behaves. Following the specification to the letter however,
the shadow wouldn't be visible through the foreground.

This PR aligns the specification with Chrome's and Firefox's behavior:
 - Render the shape, creating image A.
 - Multiply the alpha component of every pixel in A by global alpha,
   creating image B.
 - When the current filter is set, use it on image B to create image C.