Fix language stating when start tags must not be omitted
This change corrects the following statement:

> However, a start tag must never be omitted if it has any attributes.

The correction in this change makes it clear that the requirement is
about whether the start tag’s _element_ has attributes — not about
whether the start tag itself does.

Otherwise, without this change, the problem with the text as written is
the word “it”, which incorrectly refers to the start tag itself.