Disambiguate attribute declarations from content attributes
The change updates the spec to use the term “attribute declaration” in
all cases where what the spec is referring to is actually the syntax for
declaring attributes in markup.

Otherwise, without this change, the spec uses the same term “attribute”
to refer both to attribute declarations in markup and to actual content
attributes as they exist in the DOM.

And we have evidence of authors being confused due to that ambiguous use
of the same term to refer to different things — and evidence suggesting
the ambiguous usage promotes the wrong mental model of HTML for authors.