Change URL parsing sequence for window open steps
Previously, we would only check at the last minute. This would be buggy in the case where we create a new navigable, and then throw an exception. Instead we should be sure to throw before creating a new navigable.

In addition to fixing this somewhat-obviously-wrong bug, this also causes validation to happen before popup blocking or other window-choosing steps. Previously, if trying to open a window to an unparseable URL with the popup blocker active, the window open steps would return null; now, they throw an exception.

Closes #10681.