Allow dedicated Workers to inherit the creator's import map
This option adds an `importMap: "none" | "clone"` option to the `Worker`
constructor, which defaults to `"none"`. When this option is set to
`"clone"`, the `Worker` constructor will take a snapshot of the current
import map and use it as the import map inside the worker.

Passing the `importMap` option to `SharedWorker` is an error, to make it
possible to give it a meaning in the future.