Make named window lookup more precise and correct BCG swap logic
In "the rules for choosing a navigable," the method to find an existing
navigable by name is vague. This updates the definition to accurately
reflect what the major implementations do. There are some differences
between implementations, so there remains in the spec some
optional/implementation-defined behaviour, but it's much narrower.
In particular, note that lookups are now explicitly scoped to
browsing context groups. The previous language in the named lookup
about "the user agent determines that the two browsing contexts are
related enough" is now no longer a part of the lookup logic, but a
consequence of the BCG swap decisions.

In "obtain a browsing context to use for a navigation response," the
existing spec only mentions COOP enforcement as a reason to do a
browsing context group swap. Some implementations perform a swap for
additional security and performance reasons. This is now reflected
in the spec.