Fix url and src in NotRestoredReasons
url was previously listed as a DOMString, but it should be a USVString. This change is not observable since it always contained a serialized URL anyway.

src was previously listed as containing the literal value of the src="" content attribute. However, the intended behavior was to contain the parsed-then-serialized value of the src IDL attribute, i.e., the result after content/IDL attribute reflection. Fix the spec to reflect that, which allows also changing it to USVString.

Finally, also account for the fact that sometimes content attributes can be missing, in which case these properties need to return the empty string. (Not null, which is reserved for the top level or non-iframe child cases.)

Closes #10895.