Add "has storage access" boolean to environment
This is a concept that originated in the Storage Access API, where it
has been stuck because of spec issues between Fetch and 6265bis. To
un-logjam this, I've started https://github.com/whatwg/fetch/pull/1807.
That depends on this bit existing.

This patch adds the bit, which remains false, and does nothing.