Fix dialog focus restoring descendant check
In https://github.com/whatwg/html/pull/9178 I accidentally made the
currently focused element check look for the element to restore focus on
as an ancestor instead of the dialog element "subject". This PR corrects
this to look at the dialog element instead of the element to restore
focus to.

This was identified here: https://github.com/whatwg/html/pull/9178#issuecomment-2620457606