Editorial: clean up navigation API "resolve the finished promise"
Remove a misleading no-op step, and an inaccurate note, from this algorithm. Instead, replace them with an assert that such a step is unnecessary.

It's likely that in previous revisions of the specification text (perhaps before moving to HTML), this step did something. But it definitely is unnecessary now.

Fixes #11064.