Editorial: use empty <span> for old ID
This is the style used in many other places to keep old IDs working.

Doing this consistently will make it easier to detect this pattern and
transform it to oldids="" for Bikeshed, and makes it clear in the source
that the <span> isn't creating a link, as they usually do.