Clarify that "this" is not available while in parallel
Fixes #11110, by clarifying that Web IDL's this value is not available from steps running in parallel, even if those steps were kicked off my a non-parallel algorithm that _can_ access "this".

Clarify this in the parallelism example, and in prose about how to operate from while parallel, and fix all cases of "this's relevant _foo_" being used improperly from parallel steps, which is a good start to setting the right precedent.