Add ImageData Float16Array support
Add the ImageDataPixelFormat enum type, which can take on values of
"rgba-unorm8" (the current default of 8-bit unsigned normalized RGBA)
and "rgba-float16" (the new option for 16-bit floating point RGBA).

Add ImageDataPixelFormat to the ImageDataSettings dictionary used by
ImageData constructors, to allow creation of ImageData objects that are
backed by Float16Array.

Add ImageDataPixelFormat as an attribute to ImageData.

Add a new ImageDataArray type which is the union of Uint8ClampedArray
and Float16Array. Change the ImageData data attribute to be this union
type instead of Uint8ClampedArray.

Fix the Canvas Pixel ArrayBuffer section to correctly describe the pixel
layout of an ImageBuffer (there was a bug in the spec specifying that an
ImageData's pixel data starts at the beginning of its data attributes's
backing ArrayBuffer, when what was intended was for it to start at the
beginning of its data attribute).

Fixes #10856