Fix source insertion/moving/removing steps
As discovered in https://github.com/whatwg/html/pull/10657, no browser's removing or moving steps rely on the 'next sibling' pointer associated with the old element's position. After writing a test, we also discovered that the same goes for the insertion steps. In all cases, when the parent (old parent, for removing/moving steps) is a <picture> element, a relevant mutation is triggered for all <img> children of the parent. Update the spec to reflect this.

Closes #11113.