Fix document access in shared history push/replace steps
Closes #11199.

As discussed there, there are deeper issues around whether a History object is associated with a Window vs. History object, and how that intersects with #3267 around Window object reuse.

However, in the meantime, this makes the traversal from History → Document consistent with the rest of the specification, and stops using the undefined concept of "history's associated Document".