Remove cross-origin check from autofocus insertion steps
Step 6 is not needed anymore after the addition of "allow focus steps" in step 5 that checks if the element belongs to an element that allows the feature policy focus-without-user-activation. I.e., if an element is in a frame that allows the policy with *, then it should allow autofocus even if it's from a different origin.

Fixes  #11276.